[fix] actually apply new class
