+++ package.json
@@ -78,6 +78,7 @@
     "@ngrx/store-devtools": "^6.0.1",
     "@nguniversal/express-engine": "^6.0.0",
     "@nguniversal/module-map-ngfactory-loader": "^6.0.0",
+    "@types/core-js": "^2.5.0",
     "@types/jasmine": "~2.8.9",
     "@types/jasminewd2": "~2.0.5",
     "@types/node": "~10.3.4",
@@ -86,14 +87,14 @@
     "core-js": "^2.4.1",
     "express": "^4.16.3",
     "hammerjs": "^2.0.8",
-    "jasmine-core": "~2.8.0",
+    "jasmine-core": "^3.3.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^2.0.2",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.4.3",
-    "karma-jasmine": "^1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma": "^3.1.4",
+    "karma-chrome-launcher": "^2.2.0",
+    "karma-cli": "^2.0.0",
+    "karma-coverage-istanbul-reporter": "^2.0.4",
+    "karma-jasmine": "^2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
     "mocha": "^3.0.2",
     "ng-packagr": "^3.0.6",
     "node-sass": "^4.9.2",
