+++ src/app/ngEngine/ng-engine.ts
@@ -2,7 +2,7 @@
 import { Inject, Injectable, InjectionToken } from '@angular/core'
 
 // Config Class
-import { NgEngineConfig } from './ng-engine.config.old'
+import { NgEngineConfig } from './ng-engine.config'
 
 // Configuration Interface
 import { DefaultNgEngineConfiguration, NgEngineConfiguration } from './ng-engine.interface'
