+++ tsconfig.json
@@ -16,7 +16,6 @@
       "es2017",
       "dom"
     ],
-    "baseUrl": ".",
     "paths": {
       "@app/actions/*": [
         "./src/app/root/store/actions/*"
